Update comments Move toVnetPeering to helper.go
