+++ service/azureconfig/v1/resource/vnetpeering/desired.go
@@ -12,7 +12,7 @@ import (
 )
 
 // GetDesiredState return desired peering for host cluster virtual network.
-// Peering resource is named after cluster's resource group and targeting its virtual network.
+// Peering resource is named after guest cluster's resource group and targeting its virtual network.
 func (r Resource) GetDesiredState(ctx context.Context, azureConfig interface{}) (interface{}, error) {
 	a, err := key.ToCustomObject(azureConfig)
 	if err != nil {
